Refactor queryHandler to support 'document' and 'code' strategies
The queryHandler function has been refactored to include distinct processing paths depending on whether the 'document' or 'code' strategy is being used. This improves flexibility and allows for tailored handling depending on the type of input. In addition, the strategy type now determines the prefix displayed in the page content field of a document.