Refactor and limit query size in query handler
The refactoring introduces changes to the query constructor in `query.ts` to limit the number of queries to 3. This is done by modifying `req1.body.k` and `req.body.k` in the query replace method and the following condition block respectively. Pushing `queryByVectorStore` into `queries` array is also moved within the query replace method and after the condition block for better readability.